(#11944) Add Unix and Windows exec commands
This commit defines a set of methods for executing common commands on
Unix and Windows, e.g. touch, echo. This way all of logic for
performing exec commands is in one place, rather than scattered
throughout each exec test case.

On Windows, echo is a shell built-in, but there isn't a shell exec
provider, so it has to be wrapped within a cmd.exe shell. Touch is not
a built-in, but it can be faked using echo. Also, we have to use
backslashes when referencing the absolute path to cmd.exe and trying
to invoke it via ssh/cygwin. I'm not entirely sure why, but I can
reproduce the error when running a bash shell on the local Windows
box. The fix is to use backslashes for the absolute path.

This commit assumes the windows directory is in c:/windows, which
isn't ideal. It could be made more robust by querying the remote
system for its ENV['windir'] and ENV['COMSPEC'], or allowing them to
be configured, but that's more than I want to do right now.