Allow helper and load_path options to be strings in options file
Previously the harness would allow one or more helper files or additions
to the load path via command line flags or within an options file.
However the values in the file for these two options had to be arrays,
even if only one helper or additon to the load path was needed. This
allows the options in the file to be specified as either an array or a
single string.