Coerce the "tests" options into an Array
Previously we tried to iterate on each file in the "tests" option passed
to TestSuite, this assumed that the option was an array. Since what
little user validation actually happens, happens in our CLI parser when a
single directory string was passed via the options file this would blow
up. This forces the tests option into an Array within TestSuite until we
have a more structured way of performing user input validation.