Dump syslog on either rhel or debian if fail to kill puppet master
It's very difficult to debug why we were unable to kill the puppet
master (presumably it died or was killed by an earlier action).  And
usually our vms are spun down after execution.  If it's a transient
issue, then trying to track it down in a controlled environment is
difficult as well.  So this takes a stab at dumping the default logdest
target (syslog) on either a rhel or debian based system.